Reorganise Json ctors, move parser to new file
