+++ src/kernel/json.c
@@ -1,33 +1,12 @@
+#include "./json.h"
+
 #include <stdio.h>
 
-#include "./gc.h"
-#include "./types.h"
 #include "./wrapper/wrapper.h"
-
-#define CTOR_KERNEL_ARRAY 0xffffffff
-#define KERNEL_CTOR_OFFSET 1024 * 1000
-
-enum {
-  SUCCEED = KERNEL_CTOR_OFFSET,
-  FAIL,
-  INT,
-  BOOL,
-  FLOAT,
-  VALUE,
-  STRING,
-  LIST,
-  ARRAY,
-  NULL_,
-  FIELD,
-  INDEX,
-  KEY_VALUE,
-  MAP,
-  AND_THEN,
-  ONE_OF,
-};
+#include "./string.h"
 
 static void* eval_Json_succeed(void* args[]) {
-  return NEW_CUSTOM(SUCCEED, 1, args);
+  return NEW_CUSTOM(DECODER_SUCCEED, 1, args);
 }
 Closure Json_succeed = {
     .header = HEADER_CLOSURE(0),
@@ -36,7 +15,7 @@ Closure Json_succeed = {
 };
 
 static void* eval_Json_fail(void* args[]) {
-  return NEW_CUSTOM(FAIL, 1, args);
+  return NEW_CUSTOM(DECODER_FAIL, 1, args);
 }
 Closure Json_fail = {
     .header = HEADER_CLOSURE(0),
@@ -46,28 +25,28 @@ Closure Json_fail = {
 
 Custom Json_decodeInt = {
     .header = HEADER_CUSTOM(0),
-    .ctor = INT,
+    .ctor = DECODER_INT,
 };
 Custom Json_decodeBool = {
     .header = HEADER_CUSTOM(0),
-    .ctor = BOOL,
+    .ctor = DECODER_BOOL,
 };
 Custom Json_decodeFloat = {
     .header = HEADER_CUSTOM(0),
-    .ctor = FLOAT,
+    .ctor = DECODER_FLOAT,
 };
 Custom Json_decodeValue = {
     .header = HEADER_CUSTOM(0),
-    .ctor = VALUE,
+    .ctor = DECODER_VALUE,
 };
 Custom Json_decodeString = {
     .header = HEADER_CUSTOM(0),
-    .ctor = STRING,
+    .ctor = DECODER_STRING,
 };
 
 static void* eval_Json_decodeList(void* args[]) {
   void* decoder = args[0];
-  return NEW_CUSTOM(LIST, 2, ((void*[]){&Unit, decoder}));
+  return NEW_CUSTOM(DECODER_LIST, 2, ((void*[]){&Unit, decoder}));
 }
 Closure Json_decodeList = {
     .header = HEADER_CLOSURE(0),
@@ -77,7 +56,7 @@ Closure Json_decodeList = {
 
 static void* eval_Json_decodeArray(void* args[]) {
   void* decoder = args[0];
-  return NEW_CUSTOM(ARRAY, 2, ((void*[]){&Unit, decoder}));
+  return NEW_CUSTOM(DECODER_ARRAY, 2, ((void*[]){&Unit, decoder}));
 }
 Closure Json_decodeArray = {
     .header = HEADER_CLOSURE(0),
@@ -87,7 +66,7 @@ Closure Json_decodeArray = {
 
 static void* eval_Json_decodeNull(void* args[]) {
   void* value = args[0];
-  return NEW_CUSTOM(NULL_, 3, ((void*[]){&Unit, &Unit, value}));
+  return NEW_CUSTOM(DECODER_NULL, 3, ((void*[]){&Unit, &Unit, value}));
 }
 Closure Json_decodeNull = {
     .header = HEADER_CLOSURE(0),
@@ -98,7 +77,7 @@ Closure Json_decodeNull = {
 static void* eval_Json_decodeField(void* args[]) {
   void* field = args[0];
   void* decoder = args[1];
-  return NEW_CUSTOM(FIELD,
+  return NEW_CUSTOM(DECODER_FIELD,
       4,
       ((void*[]){
           /*a*/ &Unit,
@@ -116,7 +95,7 @@ Closure Json_decodeField = {
 static void* eval_Json_decodeIndex(void* args[]) {
   void* index = args[0];
   void* decoder = args[1];
-  return NEW_CUSTOM(INDEX,
+  return NEW_CUSTOM(DECODER_INDEX,
       5,
       ((void*[]){
           /*a*/ &Unit,
@@ -134,7 +113,7 @@ Closure Json_decodeIndex = {
 
 static void* eval_Json_decodeKeyValuePairs(void* args[]) {
   void* decoder = args[0];
-  return NEW_CUSTOM(KEY_VALUE, 2, ((void*[]){&Unit, decoder}));
+  return NEW_CUSTOM(DECODER_KEY_VALUE, 2, ((void*[]){&Unit, decoder}));
 }
 Closure Json_decodeKeyValuePairs = {
     .header = HEADER_CLOSURE(0),
@@ -145,7 +124,7 @@ Closure Json_decodeKeyValuePairs = {
 static void* eval_Json_andThen(void* args[]) {
   void* callback = args[0];
   void* decoder = args[1];
-  return NEW_CUSTOM(AND_THEN,
+  return NEW_CUSTOM(DECODER_AND_THEN,
       8,
       ((void*[]){
           /*a*/ &Unit,
@@ -166,7 +145,7 @@ Closure Json_andThen = {
 
 static void* eval_Json_oneOf(void* args[]) {
   void* decoders = args[0];
-  return NEW_CUSTOM(ONE_OF,
+  return NEW_CUSTOM(DECODER_ONE_OF,
       7,
       ((void*[]){
           /*a*/ &Unit,
@@ -193,8 +172,8 @@ Closure Json_oneOf = {
 static void* eval_Json_mapMany(void* args[]) {
   size_t n_decoders = (size_t)args[0];
   Closure* f = args[1];
-  Custom* decoders = NEW_CUSTOM(CTOR_KERNEL_ARRAY, (u32)n_decoders, &args[2]);
-  return NEW_CUSTOM(MAP,
+  Custom* decoders = NEW_CUSTOM(JSON_VALUE_ARRAY, (u32)n_decoders, &args[2]);
+  return NEW_CUSTOM(DECODER_MAP,
       7,
       ((void*[]){
           /*a*/ &Unit,
@@ -317,208 +296,79 @@ https://ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf
 
 */
 
-void skip_whitespace(u16** cursor, u16* end) {
-  u16* ptr;
-  for (ptr = *cursor; ptr < end; ptr++) {
-    u16 c = *ptr;
-    if (c != ' ' && c != '\t' && c != '\r' && c != '\n') {
-      break;
-    }
-  }
-  *cursor = ptr;
-}
+void* runStringHelp(Custom* decoder, u16** cursor, u16* end) {
+  // void* result;
 
-void* parse_bool(u16** cursor, u16* end) {
-  size_t len = end - *cursor;
-  u16* chars = *cursor;
-  u64* bytes64 = (u64*)chars;
+  // skip_whitespace(cursor, end);
 
-  const u64 BYTES_TRUE = 0x0065007500720074;
-  const u64 BYTES_FALS = 0x0073006c00610066;
-
-  if (len >= 4 && bytes64[0] == BYTES_TRUE) {
-    *cursor += 4;
-    return &True;
-  } else if (len >= 5 && bytes64[0] == BYTES_FALS && chars[4] == 'e') {
-    *cursor += 5;
-    return &False;
-  } else {
-    return NULL;
-  }
-}
-
-void* parse_null(u16** cursor, u16* end) {
-  size_t len = end - *cursor;
-  u64* bytes64 = (u64*)(*cursor);
-
-  const u64 BYTES_NULL = 0x006c006c0075006e;
-
-  if (len >= 4 && bytes64[0] == BYTES_NULL) {
-    *cursor += 4;
-    return &JsNull;
-  } else {
-    return NULL;
-  }
-}
-
-void* parse_int(u16** cursor, u16* end) {
-  u16* chars = *cursor;
-  i64 total = 0;
-
-  size_t negative = chars[0] == '-';
-  size_t start = negative;
-  size_t len = end - chars;
-  if (len > 11) len = 11;
-
-  size_t i;
-  for (i = start; i <= len; i++) {
-    i64 digit = chars[i] - '0';
-    if (digit < 0 || digit > 9) break;
-    total = (total * 10) + digit;
-  }
-
-  if (i == start) return NULL;  // no digits found
-  if (negative) {
-    total = -total;
-    if (total < -__INT32_MAX__ - 1) return NULL;
-  } else {
-    if (total > __INT32_MAX__) return NULL;
-  }
-
-  *cursor += i;
-  return NEW_ELM_INT((i32)total);
-}
-
-void* parse_float(u16** cursor, u16* end) {
-  u16* chars = *cursor;
-
-  size_t max = end - chars;
-  if (max > 31) max = 31;
-
-  char digits[32];
-  size_t d = 0;
-  for (size_t i = 0; i < max; i++) {
-    char c = (char)chars[i];
-    if ((c >= '0' && c <= '9') || c == '-' || c == '+' || c == 'e' || c == 'E' ||
-        c == '.') {
-      digits[d++] = c;
-    } else {
-      break;
-    }
-  }
-  if (!d) return NULL;
-  digits[d] = '\0';
-
-  f64 f;
-  int success = sscanf(digits, "%lg", &f);
-  if (!success) return NULL;
-
-  *cursor += d;
-  return NEW_ELM_FLOAT(f);
-}
-
-void* parse_string(u16** cursor, u16* end) {
-  u16* from = *cursor;
-  u16* to;
-
-  if (from >= end) return NULL;  // make sure it's safe to deref
-  if (*from != '"') return NULL;
-  from++;
-
-  size_t alloc_chunk_bytes = 4 * SIZE_UNIT;  // Grow the output string in chunks this big
-  ElmString16* str = NEW_ELM_STRING16(alloc_chunk_bytes / 2);
-  u16* str_end = &str->words16[alloc_chunk_bytes / 2];
-
-  for (to = str->words16;; to++, from++) {
-    if (from >= end) return NULL;
-    if (*from == '"') break;
-
-    if (to >= str_end) {
-      // Grow output string as needed, taking advantage of GC 'bump allocation'
-      void* more_space = GC_malloc(alloc_chunk_bytes);
-      if (more_space == pGcFull) return pGcFull;
-      str->header.size += alloc_chunk_bytes / SIZE_UNIT;
-      str_end += alloc_chunk_bytes / 2;
-      if (alloc_chunk_bytes < 1024) alloc_chunk_bytes *= 2;
-    }
-
-    if (*from == '\\') {
-      from++;
-      if (from >= end) return NULL;
-      switch (*from) {
-        case '"':
-        case '\\':
-        case '/':
-          *to = *from;
-          break;
-        case 'b':
-          *to = '\b';
-          break;
-        case 'f':
-          *to = '\f';
-          break;
-        case 'n':
-          *to = '\n';
-          break;
-        case 'r':
-          *to = '\r';
-          break;
-        case 't':
-          *to = '\t';
-          break;
-        case 'u': {
-          from++;
-          if (from + 4 >= end) return NULL;
-          u16 c = 0;
-          for (size_t i = 0; i < 4; i++) {
-            u16 h = from[i];
-            if (h >= '0' && h <= '9') {
-              c = (c * 16) + h - '0';
-            } else if (h >= 'a' && h <= 'f') {
-              c = (c * 16) + 10 + h - 'a';
-            } else if (h >= 'A' && h <= 'F') {
-              c = (c * 16) + 10 + h - 'A';
-            } else {
-              return NULL;
-            }
-          }
-          from += 3;  // we'll get another +1 in the for loop
-          *to = c;
-          break;
-        }
-        default:
-          return NULL;
-      }
-    } else if (*from > 0x1f) {
-      *to = *from;
-    } else {
-      return NULL;
-    }
-  }
-
-  // normalise the string length, chopping off any over-allocated space
-  size_t truncated_str_end_addr = ((size_t)to + SIZE_UNIT - 1) & (-SIZE_UNIT);
-  size_t final_size = (truncated_str_end_addr - (size_t)str) / SIZE_UNIT;
-  str->header.size = (u32)final_size;
-  for (; to < (u16*)truncated_str_end_addr; to++) {
-    *to = 0;
-  }
-
-  *cursor = ++from;
-
-  return str;
+  // switch (decoder->ctor) {
+  //   case DECODER_SUCCEED:
+  //     break;
+  //   case DECODER_FAIL:
+  //     break;
+  //   case DECODER_INT:
+  //     result = parse_int(cursor, end);
+  //     break;
+  //   case DECODER_BOOL:
+  //     result = parse_bool(cursor, end);
+  //     break;
+  //   case DECODER_FLOAT:
+  //     result = parse_number(cursor, end);
+  //     break;
+  //   case DECODER_VALUE:
+  //     // data structure representing the actual JSON
+  //     // ADT with 0=JSref, 1=prim, 2=array, 3=object
+  //     // wrapper should know about this at some point
+  //     // need an array parser and an object parser
+  //     break;
+  //   case DECODER_STRING:
+  //     result = parse_string(cursor, end);
+  //     break;
+  //   case DECODER_LIST:
+  //     // run the array parser & convert the Custom to a List
+  //     // use List_create!
+  //     break;
+  //   case DECODER_ARRAY:
+  //     // run the array parser & convert the Custom to an Elm Array
+  //     break;
+  //   case DECODER_NULL:
+  //     result = parse_null(cursor, end);
+  //     break;
+  //   case DECODER_FIELD:
+  //     // fields might not be in order!!
+  //     // pre-parsing would give faster access to each field
+  //     result = parse_string(cursor, end);
+  //     break;
+  //   case DECODER_INDEX:
+  //     // skip bits of the array
+  //     break;
+  //   case DECODER_KEY_VALUE:
+  //     break;
+  //   case DECODER_MAP:
+  //     break;
+  //   case DECODER_AND_THEN:
+  //     break;
+  //   case DECODER_ONE_OF:
+  //     break;
+  //   default:
+  //     break;
+  // }
+  return NULL;
 }
 
 static void* eval_runOnString(void* args[]) {
-  size_t jsIndex = (size_t)args[0];
-  return Wrapper_callJsSync(jsIndex, 2, &args[1]);
+  Custom* decoder = args[0];
+  ElmString16* string = args[1];
+
+  u16* cursor = string->words16;
+  u16* end = string->words16 + code_units(string);
+
+  return runStringHelp(decoder, &cursor, end);
 }
 Closure Json_runOnString = {
     .header = HEADER_CLOSURE(0),
-    .n_values = 1,
-    .max_values = 3,
+    .max_values = 2,
     .evaluator = &eval_runOnString,
-    .values = {NULL},
 };
 
 static void* eval_run(void* args[]) {
